pull sslconfig reading out of boot
